[~] AddPipeline
